Move derivation related traits to transformer subpackage
