+++ chimney/src/main/scala-2/io/scalaland/chimney/internal/compiletime/derivation/transformer/DerivationPlatform.scala
@@ -1,7 +1,7 @@
-package io.scalaland.chimney.internal.compiletime.derivation
+package io.scalaland.chimney.internal.compiletime.derivation.transformer
 
-import io.scalaland.chimney.{partial, PartialTransformer, Transformer}
 import io.scalaland.chimney.internal.compiletime.{DefinitionsPlatform, DerivationResult}
+import io.scalaland.chimney.{partial, PartialTransformer, Transformer}
 
 private[derivation] trait DerivationPlatform extends Derivation with Legacy { this: DefinitionsPlatform =>
