+++ chimney/src/main/scala-2/io/scalaland/chimney/internal/compiletime/derivation/transformer/GatewayPlatform.scala
@@ -1,7 +1,7 @@
-package io.scalaland.chimney.internal.compiletime.derivation
+package io.scalaland.chimney.internal.compiletime.derivation.transformer
 
-import io.scalaland.chimney.{internal, PartialTransformer, Transformer}
 import io.scalaland.chimney.internal.compiletime.DefinitionsPlatform
+import io.scalaland.chimney.{internal, PartialTransformer, Transformer}
 
 import scala.annotation.unused
 
@@ -9,7 +9,6 @@ private[compiletime] trait GatewayPlatform extends Gateway {
   this: DefinitionsPlatform & DerivationPlatform & LegacyPlatform =>
 
   import c.universe.{internal as _, Transformer as _, *}
-
   import typeUtils.fromWeakConversion.*
 
   def deriveTotalTransformerImpl[
