+++ chimney/src/main/scala/io/scalaland/chimney/internal/compiletime/derivation/transformer/Gateway.scala
@@ -1,8 +1,7 @@
-package io.scalaland.chimney.internal.compiletime.derivation
+package io.scalaland.chimney.internal.compiletime.derivation.transformer
 
-import io.scalaland.chimney.{PartialTransformer, Transformer}
-import io.scalaland.chimney.internal
 import io.scalaland.chimney.internal.compiletime.{Definitions, DerivationResult}
+import io.scalaland.chimney.{internal, PartialTransformer, Transformer}
 
 import scala.annotation.nowarn
