+++ chimney/src/main/scala/io/scalaland/chimney/internal/compiletime/derivation/transformer/Legacy.scala
@@ -1,4 +1,4 @@
-package io.scalaland.chimney.internal.compiletime.derivation
+package io.scalaland.chimney.internal.compiletime.derivation.transformer
 
 import io.scalaland.chimney.internal.compiletime.{Configurations, Contexts, Definitions, DerivationResult}
